refactor: FortuneAPIのユニットテスト用のモックプロトコルを外部ファイルに分離 プロジェクトにmdやSecrets.configを追加
